+++ src/opengl/gl_draw.cpp
@@ -38,7 +38,7 @@
 struct FontSize
 {
     Fl_Font font;
-    double size;
+    float size;
     FontSize* left, *right;
     int listbase;
 #if USE_XFT
@@ -47,7 +47,7 @@ struct FontSize
 };
 static FontSize* root, *current;
 
-void gl_font(Fl_Font font, double size)
+void gl_font(Fl_Font font, float size)
 {
     fl_font(font, size);         // necessary so fl_measure() works
     size = fl_size();            // get the rounded value
@@ -89,7 +89,7 @@ void gl_font(Fl_Font font, double size)
 }
 
 
-void gl_font(int fontid, double size)
+void gl_font(int fontid, float size)
 {
     gl_font(fl_fonts + (fontid % 16), size);
 }
@@ -101,7 +101,7 @@ void gl_draw(const char* str, int n)
 }
 
 
-void gl_draw(const char* str, int n, double x, double y, double z)
+void gl_draw(const char* str, int n, float x, float y, float z)
 {
     glRasterPos3d(x, y, z);
     gl_draw(str, n);
@@ -114,7 +114,7 @@ void gl_draw(const char* str)
 }
 
 
-void gl_draw(const char* str, double x, double y, double z)
+void gl_draw(const char* str, float x, float y, float z)
 {
     gl_draw(str, strlen(str), x, y, z);
 }
@@ -123,30 +123,30 @@ void gl_draw(const char* str, double x, double y, double z)
 #if USE_XFT
 // We must use the X font, the normal functions will use the Xft font:
 
-double gl_height()
+float gl_height()
 {
     return current->xfont->ascent+current->xfont->descent;
 }
 
 
-double gl_descent() { return current->xfont->descent; }
+float gl_descent() { return current->xfont->descent; }
 
-double gl_width(const char* s, int n)
+float gl_width(const char* s, int n)
 {
     return XTextWidth(current->xfont, s, n);
 }
 
 
-double gl_width(const char* s) {return gl_width(s, strlen(s));}
+float gl_width(const char* s) {return gl_width(s, strlen(s));}
 
 #else
 // The old X and Windows versions use exactly the same fonts for OpenGL
 // and for normal drawing, so we can share the functions:
 
-double gl_height() {return fl_height();}
-double gl_descent() {return fl_descent();}
-double gl_width(const char* s) {return fl_width(s);}
-double gl_width(const char* s, int n) {return fl_width(s,n);}
+float gl_height() {return fl_height();}
+float gl_descent() {return fl_descent();}
+float gl_width(const char* s) {return fl_width(s);}
+float gl_width(const char* s, int n) {return fl_width(s,n);}
 #endif
 
 void gl_rect(int x, int y, int w, int h)
